pyproject.toml
[tool.poetry]
name = "aws-study-proj-1"
version = "0.1.0"
description = ""
authors = ["alexlinus <alexlinus93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
boto3 = "^1.36.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
